Update XrdConnection::open() promise/future
This commit adds a thread that watches the connection in progress. It
sets the value in the promise, associated with the future returned by
Open(), to true when the WebRTC connection has been established. Now
the user does not need to query isConnected(), but rather can just
check the value in the future to know when the connection is
established.