Add support for stopping AsyncWorker
This commit adds support for stopping the AsyncWorker which monitors
the connection attempt in order to set a promise that the javascript
side can look at to know when a connection attempt completed or
failed.

Now when the user calls closeConnection(), the AsyncWorker updates the
promise accordingly and stops running. This prevents an issue where
the user quits the relay app in the middle of a connection attempt and
where the connection attempt then keeps running until timing out, even
after the window closes.

See task: BOT-1914