Add Windows video forwarder support
This also changes the use of std::cerr to use fprintf() with stderr
instead. The use of std::cerr (and cout) on Windows causes a crash
when the node binding is used with an electron app. Obviously, that
shouldn't trigger a crash, but it's simpler to switch to fprintf() for
now than debug the crash. Note that the use of std::cerr/cout does not
cause a crash when the node binding is used from node by itself (no
electron app).

See task: BOT-1867