Update XrdConnection::setVideoConfig()
The member function now takes a JS object that contains the video
config. The corresponding method in binding.ts has also been updated.

This also updates the XrdConnectionBindings callbacks in binding.ts to
use an XrdConnectionEvents enum and updates the C++ code for the
'videoConfig' event to emit an XrdVideoConfig object.

See task: BOT-1861