Don't join thread in finalizer
In the finalizer for XrdConnection's threadsafe function that is used
for emitting events, don't try to join the worker
thread. XrdConnection::closeConnection() already does this by calling
XrdConnection::stopEmitter(). Joining from the finalizer can cause the
main node.js thread to block indefinitely as there is a race between
calling closeConnection() and openConnection().

After calling closeConnection(), the threadsafe function's finalizer
gets scheduled to run on node.js's main thread. When the finalizer
runs, it trys to join _workerThread through the reference the
finalizer holds to the _workerThread object. However, openConnection()
sets _workerThread to a new object and starts a new thread. So if the
finalizer does not run until after openConnection() is called, then
the finalizer is going to be blocked trying to join the new
thread. This then blocks node.js's main thread.

This commit fixes this bug by making the finalizer do nothing.

See bug: BOT-1928