Compile typescript as part of install
This also moves dependencies that are needed to compile the typescript
from the "devDependencies" section to the "dependencies". The reason
for this is because when botlink-xrd-sdk is pulled into another
project (e.g., via yarn add), "devDependencies" don't get
installed. So the dependencies needed for compiling typescript need to
be in "dependencies".