Destroy C++ connection object upon close
v8 does not guarantee garbage collection. See
https://stackoverflow.com/a/9177949 and
https://code.google.com/archive/p/v8-juice/wikis/CommentaryOnV8.wiki#No_GC_guaranty. Instead
of letting the state machine thread from a prior instance of
Public::XrdConnection continue to run in the background, this commit
updates wrapper::XrdConnection to create an instance of
Public::XrdConnection when opening a connection and then destroy the
instance when closing a connection. This cleans up any resources used
by the C++ object, including the state machine thread.